Inserted comments into both HTML files.  One was supposed to be a cityscape (very primitive and small), and the other one was supposed to be a model of pool balls on a poll table.
As I say in the comments, a LOT OF geometric transformational math was involved in these, especially with the pool balls since I could only work with the x and z axes.

If you are interested in seeing my drawings and sketches on how I attempted the billiards model, just let me know and I will post it here on repl.it.